Refactoring by putting libavcodec code into its own file
